# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Manufacturing Analytics Platform - Agent API
  description: ISO-compliant multi-agent system for manufacturing analytics
  version: 1.0.0
  contact:
    name: API Support
    email: api@manufacturing-analytics.com
servers:
  - url: https://api.manufacturing-analytics.com/v1
    description: Production server
  - url: https://staging-api.manufacturing-analytics.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /agents/classify:
    post:
      tags: [Agents]
      summary: Classify intent
      description: Classify user input into manufacturing-specific intents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassifyRequest'
      responses:
        '200':
          description: Intent classified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassifyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ClassifyRequest:
      type: object
      required: [sessionId, input]
      properties:
        sessionId:
          type: string
          description: Unique session identifier
        input:
          type: string
          minLength: 1
          maxLength: 1000
          description: User input to classify
        context:
          type: object
          description: Additional context
    ClassifyResponse:
      type: object
      required: [intent, confidence]
      properties:
        intent:
          type: string
          description: Classified intent
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Classification confidence
        description:
          type: string
          description: Intent description
        suggestedActions:
          type: array
          items:
            type: string
          description: Suggested next actions
        requiresAuth:
          type: boolean
          description: Whether this intent requires authentication
        isoStandards:
          type: array
          items:
            type: string
          description: Relevant ISO standards
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Rate limit maximum
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests
        X-RateLimit-Reset:
          schema:
            type: string
            format: date-time
          description: Rate limit reset time
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'